Removes admin authentication from getAllUsers
