+++ backend/app/controllers/admin/users/index.js
@@ -1,2 +1,1 @@
-export { getAllUsers } from "./getAllUsers";
 export { updateUser } from "./updateUser";
