+++ backend/app/controllers/users/getAllUsers.js
@@ -1,8 +1,8 @@
-import * as adminService from "../../../services/admin/users/index";
+import * as userService from "../../services/users/index";
 
 export const getAllUsers = async (req, res, next) => {
   try {
-    const users = await adminService.getAllUsers();
+    const users = await userService.getAllUsers();
     res.status(200).json(users);
   } catch (error) {
     next(error);
