+++ backend/app/repositories/users/index.js
@@ -4,3 +4,4 @@ export { deleteUser } from "./deleteUser";
 export { addAttendee } from "./addAttendee";
 export { getAllEvents } from "./getAllEvents";
 export { deleteAttendee } from "./deleteAttendee";
+export { getAllUsers } from "./getAllUsers";
