+++ backend/app/services/users/getAllUsers.js
@@ -1,4 +1,4 @@
-import * as adminRepository from "../../../repositories/admin/users/index";
+import * as adminRepository from "../../repositories/users/index";
 
 export const getAllUsers = async () => {
   const users = await adminRepository.getAllUsers();
