+++ backend/app/services/users/index.js
@@ -3,3 +3,4 @@ export { getUserById } from "./getUserById";
 export { deleteUser } from "./deleteUser";
 export { addAttendee } from "./addAttendee";
 export { getAllEvents } from "./getAllEvents";
+export { getAllUsers } from "./getAllUsers";
